pyproject.toml
[project]
name = "write-book-template"  # TODO: replace with your actual project name
version = "0.1.0"
description = "Template project for writing and publishing books using markdown and automation tools"  # TODO: customize
authors = [
    { name = "Asterios Raptis", email = "asteri.raptis@gmail.com" }
]
readme = "README.md"

[tool.poetry]
packages = [
    { include = "scripts" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pandas = "^2.3.2"
python-dotenv = "^1.1.0"
requests = "^2.32.5"
pillow = "^11.2.1"
pyyaml = "^6.0.2"
toml = "^0.10.2"
ftfy = "^6.3.1"
questionary = "^2.1.1"
pymupdf = "^1.26.4"

[tool.poetry.group.dev.dependencies]
hypothesis = "^6.138.13"
types-pyyaml = "^6.0.12.20250915"
types-toml = "^0.10.8.20240310"

[tool.poetry.group.test.dependencies]
hypothesis = ">=6,<7"
pytest = "^8.2"
pytest-mock = "^3.10.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
init-book-project = "scripts.init_book_project:main"
full-export-book = "scripts.full_export_book:main"
export-shortcuts = "scripts.shortcuts_export:main"
update-metadata-values = "scripts.update_metadata_values:main"
convert-paths-to-absolute = "scripts.convert_to_absolute:main"
convert-paths-to-relative = "scripts.convert_to_relative:main"
print-version-build = "scripts.print_version_build:main"
generate-audiobook = "scripts.generate_audiobook:main"
reorder-chapters = "scripts.reorder_and_rename_chapters:main"
make-tag-message = "scripts.make_tag_message:main"
bulk-change-ext = "scripts.bulk_change_extension:main"
convert-images = "scripts.convert_images:main"
generate-images-deepai = "scripts.generate_images_deepai:main"
sanitize = "scripts.sanitize:main"
pandoc-batch = "scripts.pandoc_batch:main"
clean-git-cache = "scripts.clean_git_cache:main"
full-export = "scripts.full_export_book:main"
replace-emojis = "scripts.replace_emojis:main"
translate-book-deepl = "scripts.translate_book_deepl:main"
translate-book-lmstudio = "scripts.translate_book_lmstudio:main"
create-chapters = "scripts.create_chapters:main"
generate-images = "scripts.generate_images:main"
inject-images = "scripts.inject_images:main"
shortcuts-export = "scripts.shortcuts_export:main"
shortcuts-init = "scripts.shortcuts_init:main"
shortcuts-translate = "scripts.shortcuts_translate:main"
export-comic = "scripts.full_export_comic:main"
shortcuts-export-comic = "scripts.shortcuts_export_comic:main"

# 🏁 Export Comic Shortcuts
export-comic-html = "scripts.shortcuts_export_comic:export_comic_html"
export-comic-pdf = "scripts.shortcuts_export_comic:export_comic_pdf"
# 🏁 Export Shortcuts
export-epub = "scripts.shortcuts_export:export_epub"
export-epub2 = "scripts.shortcuts_export:export_epub2"
export-e = "scripts.shortcuts_export:export_epub"
export-e2 = "scripts.shortcuts_export:export_epub2"
export-epub-with-cover = "scripts.shortcuts_export:export_epub_with_cover"
export-epub2-with-cover = "scripts.shortcuts_export:export_epub2_with_cover"
export-ewc = "scripts.shortcuts_export:export_epub_with_cover"
export-e2wc = "scripts.shortcuts_export:export_epub2_with_cover"
export-markdown = "scripts.shortcuts_export:export_markdown"
export-md = "scripts.shortcuts_export:export_markdown"
export-pdf = "scripts.shortcuts_export:export_pdf"
export-p = "scripts.shortcuts_export:export_pdf"
export-docx = "scripts.shortcuts_export:export_docx"
export-d = "scripts.shortcuts_export:export_docx"
export-all-with-cover = "scripts.shortcuts_export:all_formats_with_cover"
export-print-version-epub = "scripts.shortcuts_export:export_print_version_epub"
export-pve = "scripts.shortcuts_export:export_print_version_epub"
export-print-version-paperback = "scripts.shortcuts_export:export_print_version_paperback"
export-pvp = "scripts.shortcuts_export:export_print_version_paperback"
export-print-version-hardcover = "scripts.shortcuts_export:export_print_version_hardcover"
export-pvh = "scripts.shortcuts_export:export_print_version_hardcover"
# Safe shortcuts (skip Step 1 & 4)
export-pdf-safe = "scripts.shortcuts_export:export_pdf_safe"
export-epub-safe = "scripts.shortcuts_export:export_epub_safe"
export-es = "scripts.shortcuts_export:export_epub_safe"
export-docx-safe = "scripts.shortcuts_export:export_docx_safe"
export-markdown-safe = "scripts.shortcuts_export:export_markdown_safe"
export-pvp-safe = "scripts.shortcuts_export:export_print_version_paperback_safe"
export-pvh-safe = "scripts.shortcuts_export:export_print_version_hardcover_safe"
export-pvps = "scripts.shortcuts_export:export_print_version_paperback_safe"
export-pvhs = "scripts.shortcuts_export:export_print_version_hardcover_safe"

# Init Shortcuts
update-mdv = "scripts.shortcuts_init:run_update_metadata_values"
init-bp = "scripts.shortcuts_init:run_init_book_project"

# Translation Shortcuts
# Generic CLI Translation Shortcut
translate = "scripts.shortcuts_translate:translate"
# Verbose Translation Shortcuts
translate_manuscript_to_german = "scripts.shortcuts_translate:translate_manuscript_to_german"
translate_manuscript_from_german_to_english = "scripts.shortcuts_translate:translate_manuscript_from_german_to_english"
translate_manuscript_from_english_to_spanish = "scripts.shortcuts_translate:translate_manuscript_from_english_to_spanish"
translate_manuscript_from_german_to_spanish = "scripts.shortcuts_translate:translate_manuscript_from_german_to_spanish"
# Short Translation Aliases
translate-de-es = "scripts.shortcuts_translate:translate_manuscript_from_german_to_spanish"
translate-en-es = "scripts.shortcuts_translate:translate_manuscript_from_english_to_spanish"
translate-en-de = "scripts.shortcuts_translate:translate_manuscript_to_german"
translate-de-en = "scripts.shortcuts_translate:translate_manuscript_from_german_to_english"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=scripts --cov-report=term-missing"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pandoc_batch]
root = "manuscript"
outdir = "output"
to = "epub"
metadata_file = "config/metadata.yaml"
resource_path = ["assets"]
lang = "en"
jobs = 4
verbose = true
standalone = true
test_only = false
patch_md = true
fix_inplace = false

[tool.mypy]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = false  # you can set it to true later
ignore_missing_imports = false
